continue to clean up 20F0.c

+++ src/10AD0.c
@@ -414,7 +414,7 @@ void renderAlphaBlendedTextSprite(TextRenderArg *arg0) {
     if (scaleS == -1) {
         clipOffsetX = frameEntry->width - 1;
     }
-    if ((s16)scaleT == -1) {
+    if (scaleT == -1) {
         clipOffsetY = frameEntry->height - 1;
     }
